src/test/java/com/ssssogong/issuemanager/service/ProjectServiceTest.java
package com.ssssogong.issuemanager.service;

import com.ssssogong.issuemanager.domain.Project;
import com.ssssogong.issuemanager.repository.ProjectRepository;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.Assert.assertEquals;

@RunWith(SpringRunner.class)
@SpringBootTest
@Transactional
public class ProjectServiceTest {

    @Autowired
    ProjectRepository projectRepository;
    @Autowired
    ProjectService projectService;

    @Test
    public void 프로젝트_추가() throws Exception {
        Project project = new Project();
        project.setProjectName("SE");
        project.setSubject("OOAD");

        projectService.saveProject(project);
        assertEquals(project, projectRepository.findProject(project.getId()));
    }
}
